+++ bootmgr-rs-core/src/system/variable.rs
@@ -3,7 +3,7 @@
 //! These store a value into a UEFI variable in a custom vendor namespace.
 
 use alloc::vec::Vec;
-use smallvec::{SmallVec, smallvec};
+use tinyvec::TinyVec;
 use uefi::{
     CStr16, Status, guid,
     runtime::{self, VariableAttributes, VariableVendor},
@@ -192,7 +192,7 @@ pub fn get_variable<T: UefiVariable + 'static>(
     name: &CStr16,
     vendor: Option<VariableVendor>,
 ) -> BootResult<T> {
-    let mut buf: SmallVec<[_; size_of::<u64>()]> = smallvec![0; size_of::<T>()]; // have a max capacity of u64, the largest type
+    let mut buf: TinyVec<[_; size_of::<u64>()]> = TinyVec::with_capacity(size_of::<T>()); // have a max capacity of u64, the largest type
     let vendor = vendor.unwrap_or(runtime::VariableVendor(BOOTMGR_GUID));
     RuntimeUefiVariableStorage::get_variable(name, &vendor, &mut buf)
 }